requirements.txt
celery
Cython
Django==1.11.1
django-closuretree
django-debug-toolbar
django-jinja
django-mptt
django-oauth-toolkit==0.11.0
django-python3-ldap==0.11.0
django-redis
djangorestframework==3.5.3
djangorestframework-oauth==1.1.0
drf-pdf
drfdocs
elasticsearch-dsl==5.4
fpdf
python-ldap
Pillow
psycopg2-binary
pylibmc
PyPDF2
python-dateutil
pytz
redis
reportlab
requests
WeasyPrint
xlrd
xlwt
xmltodict
django-celery-email
django-celery-beat
django-celery-results
geopy
social-auth-app-django